requirements.txt
# Core dependencies for ticketwatch
requests>=2.31.0
cloudscraper>=1.2.71
beautifulsoup4>=4.12.2
python-dateutil>=2.8.2
aiohttp>=3.9.0

# Optional for macOS notifications
# terminal-notifier (install separately with: brew install terminal-notifier)